# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/trips": {
      "get": {
        "description": "Returns all trips. Accessible by superuser, admin, and support.",
        "operationId": "TripsController_findAll",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for RBAC. Allowed values: traveler, guide, support, admin, superuser, agency",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "traveler",
                "guide",
                "support",
                "admin",
                "superuser",
                "agency"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of all trips"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "summary": "Get all trips",
        "tags": [
          "trips"
        ]
      },
      "post": {
        "operationId": "TripsController_create",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for RBAC. Allowed values: traveler, guide, support, admin, superuser, agency",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "traveler",
                "guide",
                "support",
                "admin",
                "superuser",
                "agency"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for ownership verification",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTripDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Trip created"
          },
          "400": {
            "description": "Validation error"
          }
        },
        "summary": "Create a new trip",
        "tags": [
          "trips"
        ]
      }
    },
    "/trips/{id}": {
      "get": {
        "operationId": "TripsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for RBAC. Allowed values: traveler, guide, support, admin, superuser, agency",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "traveler",
                "guide",
                "support",
                "admin",
                "superuser",
                "agency"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for ownership verification",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Trip found"
          },
          "404": {
            "description": "Trip not found"
          }
        },
        "summary": "Get trip by ID",
        "tags": [
          "trips"
        ]
      },
      "put": {
        "operationId": "TripsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for RBAC. Allowed values: traveler, guide, support, admin, superuser, agency",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "traveler",
                "guide",
                "support",
                "admin",
                "superuser",
                "agency"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for ownership verification",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTripDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Trip updated"
          },
          "404": {
            "description": "Trip not found"
          }
        },
        "summary": "Update a trip",
        "tags": [
          "trips"
        ]
      },
      "delete": {
        "operationId": "TripsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for RBAC. Allowed values: traveler, guide, support, admin, superuser, agency",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "traveler",
                "guide",
                "support",
                "admin",
                "superuser",
                "agency"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Trip deleted"
          },
          "404": {
            "description": "Trip not found"
          }
        },
        "summary": "Delete a trip",
        "tags": [
          "trips"
        ]
      }
    },
    "/trips/traveler/{travelerId}": {
      "get": {
        "description": "Returns all trips for a traveler.",
        "operationId": "TripsController_findByTraveler",
        "parameters": [
          {
            "name": "travelerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for RBAC. Allowed values: traveler, guide, support, admin, superuser, agency",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "traveler",
                "guide",
                "support",
                "admin",
                "superuser",
                "agency"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for ownership verification",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Trips for traveler"
          }
        },
        "summary": "Get trips by traveler ID",
        "tags": [
          "trips"
        ]
      }
    },
    "/trips/guide/{guideId}": {
      "get": {
        "description": "Returns all trips assigned to a guide.",
        "operationId": "TripsController_findByGuide",
        "parameters": [
          {
            "name": "guideId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for RBAC. Allowed values: traveler, guide, support, admin, superuser, agency",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "traveler",
                "guide",
                "support",
                "admin",
                "superuser",
                "agency"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for ownership verification",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Trips for guide"
          }
        },
        "summary": "Get trips by guide ID",
        "tags": [
          "trips"
        ]
      }
    },
    "/trips/agency/{agencyId}": {
      "get": {
        "description": "Returns all trips for an agency.",
        "operationId": "TripsController_findByAgency",
        "parameters": [
          {
            "name": "agencyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for RBAC. Allowed values: traveler, guide, support, admin, superuser, agency",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "traveler",
                "guide",
                "support",
                "admin",
                "superuser",
                "agency"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for ownership verification",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Trips for agency"
          }
        },
        "summary": "Get trips by agency ID",
        "tags": [
          "trips"
        ]
      }
    },
    "/trips/{id}/status": {
      "patch": {
        "description": "Updates only the status of a trip. Guides can update status of assigned trips. Travelers can update status after payment.",
        "operationId": "TripsController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for RBAC. Allowed values: traveler, guide, support, admin, superuser, agency",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "traveler",
                "guide",
                "support",
                "admin",
                "superuser",
                "agency"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for ownership verification",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTripStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Trip status updated"
          },
          "404": {
            "description": "Trip not found"
          }
        },
        "summary": "Update trip status",
        "tags": [
          "trips"
        ]
      }
    },
...
    "/itineraries/trip/{tripId}": {
      "get": {
        "description": "Returns all itinerary items for a trip.",
        "operationId": "ItinerariesController_findByTrip",
        "parameters": [
          {
            "name": "tripId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for RBAC. Allowed values: traveler, guide, support, admin, superuser, agency",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "traveler",
                "guide",
                "support",
                "admin",
                "superuser",
                "agency"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for ownership verification",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Itinerary items for trip"
          }
        },
        "summary": "Get itinerary by trip ID",
        "tags": [
          "itineraries"
        ]
      }
    },
...
    "/expenses/trip/{tripId}": {
      "get": {
        "description": "Returns all expenses for a trip.",
        "operationId": "ExpensesController_findByTrip",
        "parameters": [
          {
            "name": "tripId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for RBAC. Allowed values: traveler, guide, support, admin, superuser, agency",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "traveler",
                "guide",
                "support",
                "admin",
                "superuser",
                "agency"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for ownership verification",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Expenses for trip"
          }
        },
        "summary": "Get expenses by trip ID",
        "tags": [
          "expenses"
        ]
      }
    },
...
    "/reviews/trip/{tripId}": {
      "get": {
        "operationId": "ReviewsController_findByTrip",
        "parameters": [
          {
            "name": "tripId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for RBAC. Allowed values: traveler, guide, support, admin, superuser, agency",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "traveler",
                "guide",
                "support",
                "admin",
                "superuser",
                "agency"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reviews for trip"
          }
        },
        "summary": "Get reviews by trip ID",
        "tags": [
          "reviews"
        ]
      }
    },
...
  },
  "info": {
    "title": "GlobeSync API",
    "description": "GlobeSync Travel Platform Backend API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateTripDto": {
        "type": "object",
        "properties": {
          "destination": {
            "type": "string",
            "example": "Maldives",
            "description": "Trip destination"
          },
          "startDate": {
            "type": "string",
            "example": "2026-08-01",
            "description": "Start date (YYYY-MM-DD)"
          },
          "endDate": {
            "type": "string",
            "example": "2026-08-07",
            "description": "End date (YYYY-MM-DD)"
          },
          "budget": {
            "type": "number",
            "example": 150000,
            "description": "Trip budget"
          },
          "travelerId": {
            "type": "number",
            "example": 1,
            "description": "Traveler ID"
          },
          "agencyId": {
            "type": "number",
            "example": 1,
            "description": "Agency ID"
          },
          "guideId": {
            "type": "number",
            "example": 1,
            "description": "Guide ID (nullable)"
          },
          "packageId": {
            "type": "number",
            "example": 1,
            "description": "Package ID (optional — omit for custom trips)"
          },
          "status": {
            "type": "string",
            "example": "Planning",
            "enum": [
              "Planning",
              "Pending",
              "Confirmed",
              "Completed",
              "Cancelled"
            ]
          }
        },
        "required": [
          "destination",
          "startDate",
          "endDate",
          "budget"
        ]
      },
      "UpdateTripDto": {
        "type": "object",
        "properties": {
          "destination": {
            "type": "string",
            "example": "Maldives"
          },
          "startDate": {
            "type": "string",
            "example": "2026-08-01"
          },
          "endDate": {
            "type": "string",
            "example": "2026-08-07"
          },
          "budget": {
            "type": "number",
            "example": 150000
          },
          "travelerId": {
            "type": "number",
            "example": 1
          },
          "agencyId": {
            "type": "number",
            "example": 1
          },
          "guideId": {
            "type": "number",
            "example": 1
          },
          "packageId": {
            "type": "number",
            "example": 1
          },
          "status": {
            "type": "string",
            "example": "Confirmed",
            "enum": [
              "Planning",
              "Pending",
              "Confirmed",
              "Completed",
              "Cancelled"
            ]
          }
        }
      },
      "UpdateTripStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "Confirmed",
            "enum": [
              "Planning",
              "Pending",
              "Confirmed",
              "Completed",
              "Cancelled"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}